Refactor cmd_admins
