+++ events/telegram/cmd_utls.go
@@ -1,6 +1,7 @@
 package telegram
 
 import (
+	"log"
 	"strconv"
 	"tg_ics_useful_bot/clients/telegram"
 	"tg_ics_useful_bot/storage"
@@ -28,4 +29,29 @@ func (a chatIDExec) Exec(p *Processor, inMessage string, user *telegram.User, ch
 	message := strconv.Itoa(chat.ID)
 	mthd := sendMessageMethod
 	return &Response{message: message, method: mthd, replyMessageId: -1}, nil
+}
+
+// allUsernamesExec предоставляет метод Exec для вызова всех админов в чате.
+type allUsernamesExec string
+
+// Exec: /all - тэгает всех админов в чате.
+func (a allUsernamesExec) Exec(p *Processor, inMessage string, user *telegram.User, chat *telegram.Chat,
+	userStats *storage.DBUserStat, messageID int) (*Response, error) {
+
+	message := a.allUsernames(chat.ID, p.tg)
+	mthd := sendMessageMethod
+	return &Response{message: message, method: mthd, replyMessageId: -1}, nil
+}
+
+// allUsernames возвращает строку "@username1, @username2...".
+func (a allUsernamesExec) allUsernames(chatID int, tgClient *telegram.Client) string {
+	admins, err := tgClient.ChatAdministrators(chatID)
+	if err != nil {
+		log.Printf("can't get admins in chat #%d: %v", chatID, err)
+	}
+	result := ""
+	for _, a := range admins {
+		result += "@" + a.Username + " "
+	}
+	return result[:len(result)-1]
 }
