+++ events/telegram/process_messages.go
@@ -2,6 +2,7 @@ package telegram
 
 import (
 	"errors"
+	"log"
 	"tg_ics_useful_bot/clients/telegram"
 	"tg_ics_useful_bot/events"
 	"tg_ics_useful_bot/lib/e"
@@ -219,6 +220,20 @@ func fetchType(upd telegram.Update) events.Type {
 func (p *Processor) isAdmin(userID int) bool {
 	for _, id := range p.tg.AdminsID {
 		if userID == id {
+			return true
+		}
+	}
+	return false
+}
+
+// isChatAdmin определяет является ли пользователь админов в чате.
+func (p *Processor) isChatAdmin(user *telegram.User, chatID int) bool {
+	admins, err := p.tg.ChatAdministrators(chatID)
+	if err != nil {
+		log.Printf("can't get admins in chat #%d: %v", chatID, err)
+	}
+	for _, admin := range admins {
+		if user.ID == admin.ID {
 			return true
 		}
 	}
